senior-project/gma_web_service/requirements.txt
click==6.6
coreapi==2.0.9
coreapi-cli==1.0.6
decorator==4.0.10
dj-database-url==0.4.1
Django==1.10.3
djangorestframework==3.5.3
gunicorn==19.6.0
httpie==0.9.6
ipython==5.1.0
ipython-genutils==0.1.0
itypes==1.1.0
pexpect==4.0.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
psycopg2==2.6.2
ptyprocess==0.5.1
requests==2.12.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
uritemplate==3.0.0
wcwidth==0.1.7
whitenoise==3.2.2
